+++ VehicleTender.Web.AdminUI/Controllers/UserOperationsController.cs
@@ -20,7 +20,7 @@ namespace VehicleTender.Web.AdminUI.Controllers
         [HttpGet]
         public async Task<IActionResult> UserList()
         {
-            if (_httpContextAccessor.HttpContext.Request.Cookies["deger"] == null)
+            if (_httpContextAccessor.HttpContext.Request.Cookies["token"] == null)
             {
                 return RedirectToAction("Login", "Auth");
             }
@@ -42,6 +42,10 @@ namespace VehicleTender.Web.AdminUI.Controllers
         [HttpGet]
         public async Task<IActionResult> UpdateUser(int id)
         {
+            if (_httpContextAccessor.HttpContext.Request.Cookies["token"] == null)
+            {
+                return RedirectToAction("Login", "Auth");
+            }
             UpdateAdmin updateAdmin = new UpdateAdmin();
             updateAdmin.isActive = false;
             updateAdmin.Telephone = "345345";
@@ -58,13 +62,12 @@ namespace VehicleTender.Web.AdminUI.Controllers
         [HttpPost]
         public async Task<IActionResult> UpdateUser(UpdateAdmin updateAdmin)
         {
-            if (ValidatorProperties.GetValidatorResult<UpdateAdmin>(updateAdmin).Count != 0)
-                Console.WriteLine("aethaeth");
-
-            if (_httpContextAccessor.HttpContext.Request.Cookies["deger"] == null)
+            if (_httpContextAccessor.HttpContext.Request.Cookies["token"] == null)
             {
                 return RedirectToAction("Login", "Auth");
             }
+            if (ValidatorProperties.GetValidatorResult<UpdateAdmin>(updateAdmin).Count != 0)
+                Console.WriteLine("aethaeth");
             var a = 4;
             Console.WriteLine(admin.getAdminDTO);
             adminService.AddNewAdmin(token, admin.addAdminDTO);
@@ -74,12 +77,13 @@ namespace VehicleTender.Web.AdminUI.Controllers
         [HttpPost]
         public async Task<IActionResult> AddNewAdmin(AdminsPage admin)
         {
-            if (ValidatorProperties.GetValidatorResult<AddAdminDTO>(admin.addAdminDTO.Username).Count != 0)
-                await adminService.AddNewAdmin(admin.addAdminDTO);
-            if (_httpContextAccessor.HttpContext.Request.Cookies["deger"] == null)
+            if (_httpContextAccessor.HttpContext.Request.Cookies["token"] == null)
             {
                 return RedirectToAction("Login", "Auth");
             }
+            if (ValidatorProperties.GetValidatorResult<AddAdminDTO>(admin.addAdminDTO.Username).Count != 0)
+                await adminService.AddNewAdmin(admin.addAdminDTO);
+
             var a = 4;
             Console.WriteLine(id);
             adminService.DeleteAdmin(token, id);
@@ -89,6 +93,10 @@ namespace VehicleTender.Web.AdminUI.Controllers
         [HttpPost]
         public IActionResult DeleteAdmin(int id)
         {
+            if (_httpContextAccessor.HttpContext.Request.Cookies["token"] == null)
+            {
+                return RedirectToAction("Login", "Auth");
+            }
             var a = 4;
             Console.WriteLine(id);
             adminService.DeleteAdmin(token, id);
